requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
authlib==1.2.1
slowapi==0.1.9

# Email support (3.0.2 compatible with Python 3.9)
aiosmtplib==3.0.2

# Production deployment dependencies
boto3==1.34.0
redis==5.0.1
rq==1.15.1
websockets==12.0
requests==2.31.0

# Async libraries for better performance
aiohttp==3.9.1
aioboto3==12.3.0

# Additional utilities
psutil==5.9.6
